finished countries

+++ mdm/infra/repository/models/country.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, String
+from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, String, BigInteger
 from sqlalchemy.orm import relationship
 
 from config.db.alembic.base import Base
@@ -13,7 +13,7 @@ class Country(Base):
     country_name = Column(String, nullable=False)
     numeric_code = Column(Integer, nullable=False)
     capital_city = Column(String, nullable=False)
-    population = Column(Integer, nullable=False)
+    population = Column(BigInteger, nullable=False)
     area = Column(Float, nullable=False)
     created_at = Column(DateTime, default=datetime.utcnow)
     updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
